requirements.txt
datasets==3.6.0
torch>=2.1.0
transformers>=4.40.0
pyarrow>=14.0.0
pillow>=10.0.0
openai-whisper>=20240930
numpy>=1.24.0
swanlab>=0.6.0
gradio>=4.44.0
